Chore: #4 - 파일, 프로토콜, 클래스명 변경 및 코드 정렬
