+++ iTunesStore/Domain/UseCases/SearchUseCase.swift
@@ -12,10 +12,10 @@ protocol SearchUseCaseInterface {
 }
 
 final class SearchUseCase: SearchUseCaseInterface {
-  
+
   private let movieRepository: MovieRepositoryInterface
   private let podcastRepository: PodcastRepositoryInterface
-  
+
   init(
     movieRepository: MovieRepositoryInterface,
     podcastRepository: PodcastRepositoryInterface
@@ -23,11 +23,12 @@ final class SearchUseCase: SearchUseCaseInterface {
     self.movieRepository = movieRepository
     self.podcastRepository = podcastRepository
   }
-  
+
   func execute(term: String) -> Single<SearchResult> {
     return Single.zip(
       movieRepository.searchMovies(term: term),
-      podcastRepository.searchPodcast(term: term))
+      podcastRepository.searchPodcast(term: term)
+    )
     .map { movies, podcasts in
       SearchResult(
         movies: movies,
